% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{barPlots}
\alias{barPlots}
\title{This function develops bar plots of frequency of all Qualitative Variables}
\usage{
barPlots(lsVS)
}
\arguments{
\item{lsVS}{is a list generated by verticalSplit function in this same package}
}
\value{
a list of bar plots
}
\description{
This function develops bar plots of frequency of all Qualitative Variables
}
\details{
The input to this function is output of verticalSplit function in this package
It outputs a list containing bar plots of frequency of all qualitative variables. 
Output must be stored in a variable.
And you could access each single bar plots as outputList[1] or outputList[2]
}
\author{
Saurabh
}

