% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{CoefV}
\alias{CoefV}
\title{Calculating coefficient of variance}
\usage{
CoefV(var)
}
\arguments{
\item{var}{any numeric vector}
}
\value{
Numeric Coefficient of variance
}
\description{
This function calculates Coefficient of variance Quantitative Variables
}
\details{
Its an internal function
}
\author{
Saurabh
}

