% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{covarianceSPlot}
\alias{covarianceSPlot}
\title{Generating ScatterPlot of coefficient of variance of all quantitative variables}
\usage{
covarianceSPlot(quantSummary)
}
\arguments{
\item{quantSummary}{is output of quantVarSumm function from explore package}
}
\value{
a Scatter Plot
}
\description{
This function generates ScatterPlot of coefficient of variance.
}
\details{
The input to this function is output of verticalSplit function in this package
It outputs a list containing pie chart of frequency of all qualitative variables. 
Output must be stored in a variable.
And you could access each single pie chart as outputList[1] or outputList[2]
}
\author{
Saurabh Jaju
}

