% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{facsummary}
\alias{facsummary}
\title{Univariate exploration of all 
Qualitative Variables in a Dataset}
\usage{
facsummary(df)
}
\arguments{
\item{df}{a data frame (your dataset)}
}
\value{
List of frequency of all internal values of all qualitative values
}
\description{
This function calculates the frequency of various qualitative values
in the data
}
\details{
Its an internal function
}
\author{
Saurabh
}

