% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{freqPiePlots}
\alias{freqPiePlots}
\title{Generating PieCharts 
The function first divides the dataset in managable smaller dataframes with \code{verticalSplit}
This function generates PieCharts of frequency of all unique values qualitative variables with \code{piePlots}}
\usage{
freqPiePlots(df)
}
\arguments{
\item{df}{your dataframe}
}
\value{
a list of PieCharts
}
\description{
Generating PieCharts 
The function first divides the dataset in managable smaller dataframes with \code{verticalSplit}
This function generates PieCharts of frequency of all unique values qualitative variables with \code{piePlots}
}
\details{
It outputs a list containing PieCharts of frequency of all qualitative variables. 
Output must be stored in a variable.
And you could access each single pie chart as 
outputList<-freqPiePlots(df)
outputList[1] or outputList[2]....
}
\author{
Saurabh Jaju
}

