% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mysummary}
\alias{mysummary}
\title{Univariate exploration of all 
Quantitative Variables in a Dataset}
\usage{
mysummary(df)
}
\arguments{
\item{df}{a data frame (your dataset)}
}
\value{
rbound values, can be efficently converted to data frame
}
\description{
This function calculates various statistics of Quantitative variables in your data
The function calculates Coef. of variance,min,max,standard deviation,variance, 
various qunatiles, mean,median,skewness,kurtosis it also checks if data contains NA's and number of unique values.
}
\details{
Its an internal function
}
\author{
Saurabh
}

