% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{qualiVarSumm}
\alias{qualiVarSumm}
\title{Generates a detailed univariate frequency summary of qualitative features in input dataset}
\usage{
qualiVarSumm(df)
}
\arguments{
\item{df}{your data frame}
}
\value{
list of variable name, unique entry, count(frequency)
}
\description{
The function first divides the dataset in managable smaller dataframes with \code{verticalSplit}
This function provides frequency count of each unique entry in each qualitative feature with \code{runFacsummary}
}
\details{
The function first divides the dataset in managable smaller dataframes
Then calculates frequency of all unique values in all qualitative features 
And returns it in the form of a list
}
\author{
Saurabh Jaju
}

