% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{quantVarSumm}
\alias{quantVarSumm}
\title{Generates a detailed univariate statistical summary of quantitative data}
\usage{
quantVarSumm(df)
}
\arguments{
\item{df}{your data frame}
}
\value{
rbound vectors of input data statistics
}
\description{
This function calculates various statistics of the quantitative features of input dataset
}
\details{
The function first divides the dataset in managable smaller dataframes with \code{verticalSplit}
Then calculates summary statistics like
Coef. of variance,minimum,maximum,standard deviation,variance, 
various qunatiles, mean,median,skewness,kurtosis with \code{runSummary}
it also checks if data contains NA's and number of unique values.
}
\author{
Saurabh Jaju
}

