% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{runFacsummary}
\alias{runFacsummary}
\title{Executes facsummary function on complete list output by verticalSplit function
 
This function calculates the frequency of various qualitative values
in the data}
\usage{
runFacsummary(lsVS)
}
\arguments{
\item{lsVS}{a list of dataframes produced by verticalSplit function}
}
\value{
List of frequency of all internal values of all qualitative values
}
\description{
Executes facsummary function on complete list output by verticalSplit function
 
This function calculates the frequency of various qualitative values
in the data
}
\details{
Its an internal function
}
\author{
Saurabh
}

