% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{runSummary}
\alias{runSummary}
\title{Executes mysummary function on complete list output by verticalSplit function}
\usage{
runSummary(lsVS)
}
\arguments{
\item{lsVS}{a list of dataframes produced by verticalSplit function}
}
\value{
rbound values, can be efficently converted to data frame
}
\description{
This function calculates various statistics of Quantitative variables in your data
The function calculates Coef. of variance,min,max,standard deviation,variance, 
various qunatiles, mean,median,skewness,kurtosis it also checks if data contains NA's and number of unique values.
}
\details{
Its an internal function
}
\author{
Saurabh
}

