% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{verticalSplit}
\alias{verticalSplit}
\title{Splitting of large datasets to more manageable data frames}
\usage{
verticalSplit(df)
}
\arguments{
\item{df}{a data frame (your dataset)}
}
\value{
List of data frames created by dividing orignal data
}
\description{
This function splits a large dataset column wise in more managable chunks of data frames 
and returns the resulting data frames in form of a list of data frames
}
\details{
Its an internal function
}
\author{
Saurabh
}

